Renamed files with plural names to singular names
